requirements.txt
Django==4.2.5
opencv-python==4.8.0.76
deepface==0.0.79
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3
opencv-contrib-python==4.8.0.76
matplotlib==3.7.2
